+++ tests/torchtune/generation/__init__.py
@@ -3,9 +3,3 @@
 #
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
-
-__version__ = ""
-
-from torchtune import datasets, models, modules, utils
-
-__all__ = [datasets, models, modules, utils]
